extract getSelectedColumns() method
